% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_leptos_dynamic_items_input.R
\name{update_shiny_leptos_dynamic_items_input}
\alias{update_shiny_leptos_dynamic_items_input}
\title{Update Dynamic List Input Widget}
\usage{
update_shiny_leptos_dynamic_items_input(
  inputId,
  value,
  session = shiny::getDefaultReactiveDomain()
)
}
\arguments{
\item{inputId}{The ID of the input object.}

\item{value}{The new value for the list. Should be a character vector or list.}

\item{session}{The Shiny session object, defaults to the current reactive domain.}
}
\description{
Changes the value of a dynamic list input widget on the client side.
}
